# TBS: Optimize for ReadTraceEvents miss